+++ src/speclj/core.clj
@@ -3,7 +3,7 @@
     [speclj.running :only (submit-description run-and-report)]
     [speclj.reporting :only (report-message)]
     [speclj.tags :only (describe-filter)]
-    [speclj.config :only (active-reporter active-runner default-runner config-mappings default-config)]
+    [speclj.config :only (active-reporters active-runner default-runner config-mappings default-config)]
     [speclj.components]
     [speclj.util :only (endl)])
   (:require
@@ -216,5 +216,5 @@ are evaluated by evaluation the file as a script.  Optional configuration paramt
           config (merge (dissoc default-config :runner) config)]
       (with-bindings (config-mappings config)
         (if-let [filter-msg (describe-filter)]
-          (report-message (active-reporter) filter-msg))
-        (run-and-report (active-runner) (active-reporter))))))
+          (report-message (active-reporters) filter-msg))
+        (run-and-report (active-runner) (active-reporters))))))