+++ src/speclj/main.clj
@@ -18,26 +18,37 @@
   (.addSwitchOption "a" "autotest" "Alias to use the 'vigilant' runner and 'specdoc' reporter.")
   (.addSwitchOption "b" "stacktrace" "Output full stacktrace")
   (.addSwitchOption "c" "color" "Show colored (red/green) output.")
-  (.addValueOption "f" "reporter" "REPORTER" (str "Specifies how to report spec results. Ouput will be written to *out*. Builtin reporters:" endl
+  (.addMultiOption "f" "reporter" "REPORTER" (str "Specifies how to report spec results. Ouput will be written to *out*. Multiple reporters are allowed.  Builtin reporters:" endl
     "  [d]ocumentation:  (description/context and characteristic names)" endl
     "  [p]rogress:       (default - dots)" endl
     "  [s]ilent:         (no output)" endl))
-  (.addValueOption "f" "format" "FORMAT" "An alias for reporter.")
-  (.addValueOption "r" "runner" "RUNNER" (str "Use a custom Runner. Builtin runners:" endl
+  (.addMultiOption "f" "format" "FORMAT" "An alias for reporter.")
+  (.addValueOption "r" "runner" "RUNNER" (str "Specifies the spec runner.  Builtin runners:" endl
     "  [s]tandard:  (default) Runs all the specs once" endl
     "  [v]igilant:  Watches for file changes and re-runs affected specs (used by autotest)" endl))
-  (.addMultiOption "t" "tag" "TAG" "Run only the characteristics with the specified tag.\nTo exclude characteristics, prefix the tag with ~ (eg ~slow).  Use this option multiple times to filter multiple tags.")
+  (.addMultiOption "t" "tag" "TAG" "Run only the characteristics with the specified tag(s).\nTo exclude characteristics, prefix the tag with ~ (eg ~slow).  Use this option multiple times to filter multiple tags.")
   (.addSwitchOption "v" "version" "Shows the current speclj version.")
   (.addSwitchOption "h" "help" "You're looking at it.")
   )
 
+(defn- resolve-runner-alias [name]
+  (cond
+    (= "s" name) "standard"
+    (= "v" name) "vigilant"
+    :else name))
+
+(defn- resolve-reporter-alias [name]
+  (cond
+    (= "d" name) "documentation"
+    (= "p" name) "progress"
+    (= "s" name) "silent"
+    :else name))
+
 (defn- resolve-aliases [options]
   (cond
     (:format options) (recur (dissoc (assoc options :reporter (:format options)) :format))
-    (:autotest options) (recur (dissoc (assoc options :runner "vigilant" :reporter "documentation") :autotest))
-    (= "d" (:reporter options)) (recur (assoc options :reporter "documentation"))
-    (= "p" (:reporter options)) (recur (assoc options :reporter "progress"))
-    (= "s" (:reporter options)) (recur (assoc options :reporter "silent"))
+    (:autotest options) (recur (dissoc (assoc options :runner "vigilant" :reporter ["documentation"]) :autotest))
+    (:reporter options) (recur (dissoc (assoc options :reporters (map resolve-reporter-alias (:reporter options))) :reporter))
     (= "s" (:runner options)) (recur (assoc options :runner "standard"))
     (= "v" (:runner options)) (recur (assoc options :runner "vigilant"))
     (:tag options) (recur (dissoc (assoc options :tags (:tag options)) :tag))
@@ -77,8 +88,8 @@
   (with-bindings (config-mappings config)
     (try
       (if-let [filter-msg (describe-filter)]
-        (report-message *reporter* filter-msg))
-      (exit (run-directories *runner* *specs* *reporter*))
+        (report-message *reporters* filter-msg))
+      (exit (run-directories *runner* *specs* *reporters*))
       (catch Exception e
         (.printStackTrace e)
         (print-stack-trace e *err*)