+++ src/speclj/running.clj
@@ -1,10 +1,10 @@
 (ns speclj.running
   (:use
     [speclj.exec :only (pass-result fail-result pending-result)]
-    [speclj.reporting :only (report-runs report-run report-description)]
+    [speclj.reporting :only (report-runs* report-run report-description*)]
     [speclj.components :only (reset-with)]
     [speclj.util :only (secs-since)]
-    [speclj.config :only (*runner* active-reporter)]
+    [speclj.config :only (*runner* active-reporters)]
     [speclj.tags :only (tags-for tag-sets-for pass-tag-filter? context-with-tags-seq)])
   (:import
     [speclj SpecPending]
@@ -34,13 +34,13 @@
       (recur @(.parent description) (concat (getter description) components))
       components)))
 
-(defn- report-result [result-constructor characteristic start-time reporter failure]
+(defn- report-result [result-constructor characteristic start-time reporters failure]
   (let [present-args (filter identity [characteristic (secs-since start-time) failure])
         result (apply result-constructor present-args)]
-    (report-run result reporter)
+    (report-run result reporters)
     result))
 
-(defn- do-characteristic [characteristic reporter]
+(defn- do-characteristic [characteristic reporters]
   (let [description @(.parent characteristic)
         befores (collect-components #(deref (.befores %)) description)
         afters (collect-components #(deref (.afters %)) description)
@@ -53,18 +53,18 @@
     (try
       (do
         (full-body)
-        (report-result pass-result characteristic start-time reporter nil))
+        (report-result pass-result characteristic start-time reporters nil))
       (catch SpecPending p
-        (report-result pending-result characteristic start-time reporter p))
+        (report-result pending-result characteristic start-time reporters p))
       (catch Exception e
-        (report-result fail-result characteristic start-time reporter e))
+        (report-result fail-result characteristic start-time reporters e))
       (finally
         (reset-withs withs))))) ;MDM - Possible clojure bug.  Inlining reset-withs results in compile error 
 
-(defn- do-characteristics [characteristics reporter]
+(defn- do-characteristics [characteristics reporters]
   (doall
     (for [characteristic characteristics]
-      (do-characteristic characteristic reporter))))
+      (do-characteristic characteristic reporters))))
 
 (defn- withs-mapping [description]
   (let [withs (concat @(.withs description) @(.with-alls description))
@@ -73,36 +73,36 @@
 
 (declare do-description)
 
-(defn- do-child-contexts [context results reporter]
+(defn- do-child-contexts [context results reporters]
   (loop [results results contexts @(.children context)]
     (if (seq contexts)
-      (recur (concat results (do-description (first contexts) reporter)) (rest contexts))
+      (recur (concat results (do-description (first contexts) reporters)) (rest contexts))
       (do
         (eval-components @(.after-alls context))
         results))))
 
-(defn- results-for-context [context reporter]
+(defn- results-for-context [context reporters]
   (if (pass-tag-filter? (tags-for context))
-    (do-characteristics @(.charcteristics context) reporter)
+    (do-characteristics @(.charcteristics context) reporters)
     []))
 
-(defn do-description [description reporter]
+(defn do-description [description reporters]
   (let [tag-sets (tag-sets-for description)]
     (when (some pass-tag-filter? tag-sets)
-      (report-description reporter description)
+      (report-description* reporters description)
       (eval-components @(.before-alls description))
       (with-bindings (withs-mapping description)
         (try
-          (let [results (results-for-context description reporter)]
-            (do-child-contexts description results reporter))
+          (let [results (results-for-context description reporters)]
+            (do-child-contexts description results reporters))
           (finally
             (reset-withs @(.with-alls description))))))))
 
 (defprotocol Runner
-  (run-directories [this directories reporter])
+  (run-directories [this directories reporters])
   (submit-description [this description])
-  (run-description [this description reporter])
-  (run-and-report [this reporter]))
+  (run-description [this description reporters])
+  (run-and-report [this reporters]))
 
 (def clj-file-regex #".*\.clj")
 (defn clj-files-in [& dirs]